test: move worker nested-exception cases to test/regression/issue/28648
worker_threads.test.ts has an unrelated pre-existing 'eval does not
leak source code' timeout under the debug+ASAN build (the 100MB x 6
fixture takes ~27s there). Moving the four new subprocess cases to the
regression file keeps that timeout out of this change's gate run.
